+++ src/main/java/it/pagopa/pagopa/apiconfig/cosmos/container/CdiCosmos.java
@@ -1,18 +1,17 @@
-package it.pagopa.pagopa.apiconfig.entity;
+package it.pagopa.pagopa.apiconfig.cosmos.container;
 
 import com.azure.spring.data.cosmos.core.mapping.Container;
 import com.azure.spring.data.cosmos.core.mapping.PartitionKey;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import java.time.LocalDateTime;
+import java.util.List;
+import javax.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-
-import javax.persistence.Id;
-import java.time.LocalDateTime;
-import java.util.List;
 
 
 @Getter
@@ -21,7 +20,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder(toBuilder = true)
-@Container(containerName = "cdis")
+@Container(containerName = "cdis",autoCreateContainer = false)
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class CdiCosmos {
