+++ src/main/java/it/pagopa/pagopa/apiconfig/cosmos/container/CdiDetailCosmos.java
@@ -1,7 +1,8 @@
-package it.pagopa.pagopa.apiconfig.entity;
+package it.pagopa.pagopa.apiconfig.cosmos.container;
 
 import com.azure.spring.data.cosmos.core.mapping.Container;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import it.pagopa.pagopa.apiconfig.entity.ServiceAmountCosmos;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,7 +18,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @SuperBuilder(toBuilder = true)
-@Container(containerName = "cdidetail")
+@Container(containerName = "cdidetail", autoCreateContainer = false)
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class CdiDetailCosmos {
     private String idBrokerPsp;
