+++ src/main/java/it/pagopa/pagopa/apiconfig/cosmos/repository/CdiCosmosRepository.java
@@ -1,7 +1,7 @@
-package it.pagopa.pagopa.apiconfig.repository;
+package it.pagopa.pagopa.apiconfig.cosmos.repository;
 
 import com.azure.spring.data.cosmos.repository.CosmosRepository;
-import it.pagopa.pagopa.apiconfig.entity.CdiCosmos;
+import it.pagopa.pagopa.apiconfig.cosmos.container.CdiCosmos;
 import org.springframework.stereotype.Repository;
 
 @Repository
