+++ src/main/java/it/pagopa/pagopa/apiconfig/cosmos/repository/PaymentTypesCosmosRepository.java
@@ -1,7 +1,7 @@
-package it.pagopa.pagopa.apiconfig.repository;
+package it.pagopa.pagopa.apiconfig.cosmos.repository;
 
 import com.azure.spring.data.cosmos.repository.CosmosRepository;
-import it.pagopa.pagopa.apiconfig.entity.PaymentTypesCosmos;
+import it.pagopa.pagopa.apiconfig.cosmos.container.PaymentTypesCosmos;
 import org.springframework.stereotype.Repository;
 
 @Repository
